Seperate groebnertools from distributedpolys
The old groebnertools.py is now groebnertools.py, which consists of
the Groebner basis algorithms, and distributedpolys.py, which
consists of the functions for manipulating sparse distributed
polynomials.
Furthermore: The tests have been split into test_groebnertools.py
and tests_distributedpolys.py, sdp_groebner is now a wrapper
around `buchberger` (the old sdp_groebner) and other algorithms
(in the future). This can be changed via the `setup` function from
polyconfig.py (I believe this is currently best, as query has to
be invoked only once in sdp_groebner and not additionally in
sdp_lcm and the tests as well).
