+++ sympy/polys/polyconfig.py
@@ -15,6 +15,8 @@ _default_config = {
 
     'GF_IRRED_METHOD'           : 'rabin',
     'GF_FACTOR_METHOD'          : 'zassenhaus',
+
+    'GB_METHOD'                 : 'buchberger',
 }
 
 _current_config = {}
